simplify structure by moving error handling out of the sync function
